+++ OpenOversight/app/auth/views.py
@@ -14,7 +14,7 @@ from flask_login import current_user, login_required, login_user, logout_user
 from .. import sitemap
 from ..email import send_email
 from ..models import User, db
-from ..utils import (
+from OpenOversight.app.util.utils import (
     HTTP_METHOD_GET,
     HTTP_METHOD_POST,
     set_dynamic_default,
