+++ OpenOversight/app/main/views.py
@@ -42,7 +42,7 @@ from ..models import (
     User,
     db,
 )
-from ..utils import (
+from OpenOversight.app.util.utils import (
     HTTP_METHOD_GET,
     HTTP_METHOD_POST,
     ac_can_edit_officer,
