+++ OpenOversight/tests/conftest.py
@@ -20,7 +20,7 @@ from xvfbwrapper import Xvfb
 from OpenOversight.app import create_app, models
 from OpenOversight.app.models import Job, Officer, Unit
 from OpenOversight.app.models import db as _db
-from OpenOversight.app.utils import ENCODING_UTF_8, merge_dicts
+from OpenOversight.app.util.utils import ENCODING_UTF_8, merge_dicts
 from OpenOversight.tests.routes.route_helpers import ADMIN_EMAIL, ADMIN_PASSWORD
