+++ OpenOversight/tests/test_commands.py
@@ -27,7 +27,7 @@ from OpenOversight.app.models import (
     Salary,
     Unit,
 )
-from OpenOversight.app.utils import get_officer
+from OpenOversight.app.util.utils import get_officer
 from OpenOversight.tests.conftest import RANK_CHOICES_1, generate_officer
