+++ OpenOversight/tests/test_utils.py
@@ -7,7 +7,7 @@ from mock import MagicMock, Mock, patch
 
 import OpenOversight
 from OpenOversight.app.models import Department, Image, Officer, Unit
-from OpenOversight.app.utils import (
+from OpenOversight.app.util.utils import (
     crop_image,
     filter_by_form,
     upload_image_to_s3_and_store_in_db,
@@ -26,7 +26,7 @@ upload_s3_patch = patch(
 
 def test_department_filter(mockdata):
     department = OpenOversight.app.models.Department.query.first()
-    results = OpenOversight.app.utils.grab_officers(
+    results = OpenOversight.app.util.utils.grab_officers(
         {
             "race": ["Not Sure"],
             "gender": ["Not Sure"],
@@ -45,7 +45,7 @@ def test_department_filter(mockdata):
 
 def test_race_filter_select_all_black_officers(mockdata):
     department = OpenOversight.app.models.Department.query.first()
-    results = OpenOversight.app.utils.grab_officers(
+    results = OpenOversight.app.util.utils.grab_officers(
         {"race": ["BLACK"], "dept": department}
     )
     for element in results.all():
@@ -54,7 +54,7 @@ def test_race_filter_select_all_black_officers(mockdata):
 
 def test_gender_filter_select_all_male_or_not_sure_officers(mockdata):
     department = OpenOversight.app.models.Department.query.first()
-    results = OpenOversight.app.utils.grab_officers(
+    results = OpenOversight.app.util.utils.grab_officers(
         {"gender": ["M"], "dept": department}
     )
 
@@ -67,7 +67,7 @@ def test_gender_filter_select_all_male_or_not_sure_officers(mockdata):
 
 def test_gender_filter_include_all_genders_if_not_sure(mockdata):
     department = OpenOversight.app.models.Department.query.first()
-    results = OpenOversight.app.utils.grab_officers(
+    results = OpenOversight.app.util.utils.grab_officers(
         {"gender": ["Not Sure"], "dept": department}
     )
 
@@ -81,7 +81,7 @@ def test_gender_filter_include_all_genders_if_not_sure(mockdata):
 
 def test_rank_filter_select_all_commanders(mockdata):
     department = OpenOversight.app.models.Department.query.first()
-    results = OpenOversight.app.utils.grab_officers(
+    results = OpenOversight.app.util.utils.grab_officers(
         {"rank": ["Commander"], "dept": department}
     )
     for element in results.all():
@@ -91,7 +91,7 @@ def test_rank_filter_select_all_commanders(mockdata):
 
 def test_rank_filter_select_all_police_officers(mockdata):
     department = OpenOversight.app.models.Department.query.first()
-    results = OpenOversight.app.utils.grab_officers(
+    results = OpenOversight.app.util.utils.grab_officers(
         {"rank": ["Police Officer"], "dept": department}
     )
     for element in results.all():
@@ -101,7 +101,7 @@ def test_rank_filter_select_all_police_officers(mockdata):
 
 def test_filter_by_name(mockdata):
     department = OpenOversight.app.models.Department.query.first()
-    results = OpenOversight.app.utils.grab_officers(
+    results = OpenOversight.app.util.utils.grab_officers(
         {"last_name": "J", "dept": department}
     )
     for element in results.all():
@@ -113,13 +113,13 @@ def test_filters_do_not_exclude_officers_without_assignments(mockdata):
     officer = OpenOversight.app.models.Officer(
         first_name="Rachel", last_name="S", department=department, birth_year=1992
     )
-    results = OpenOversight.app.utils.grab_officers({"name": "S", "dept": department})
+    results = OpenOversight.app.util.utils.grab_officers({"name": "S", "dept": department})
     assert officer in results.all()
 
 
 def test_filter_by_badge_no(mockdata):
     department = OpenOversight.app.models.Department.query.first()
-    results = OpenOversight.app.utils.grab_officers({"badge": "12", "dept": department})
+    results = OpenOversight.app.util.utils.grab_officers({"badge": "12", "dept": department})
     for element in results.all():
         assignment = element.assignments.first()
         assert "12" in str(assignment.star_no)
@@ -130,7 +130,7 @@ def test_filter_by_full_unique_internal_identifier_returns_officers(mockdata):
     target_unique_internal_id = (
         OpenOversight.app.models.Officer.query.first().unique_internal_identifier
     )
-    results = OpenOversight.app.utils.grab_officers(
+    results = OpenOversight.app.util.utils.grab_officers(
         {
             "race": ["Not Sure"],
             "gender": ["Not Sure"],
@@ -154,7 +154,7 @@ def test_filter_by_partial_unique_internal_identifier_returns_officers(mockdata)
         OpenOversight.app.models.Officer.query.first().unique_internal_identifier
     )
     partial_identifier = identifier[: len(identifier) // 2]
-    results = OpenOversight.app.utils.grab_officers(
+    results = OpenOversight.app.util.utils.grab_officers(
         {
             "race": ["Not Sure"],
             "gender": ["Not Sure"],
@@ -173,7 +173,7 @@ def test_filter_by_partial_unique_internal_identifier_returns_officers(mockdata)
 
 
 def test_compute_hash(mockdata):
-    hash_result = OpenOversight.app.utils.compute_hash(b"bacon")
+    hash_result = OpenOversight.app.util.utils.compute_hash(b"bacon")
     expected_hash = "9cca0703342e24806a9f64e08c053dca7f2cd90f10529af8ea872afb0a0c77d4"
     assert hash_result == expected_hash
 
@@ -183,7 +183,7 @@ def test_s3_upload_png(mockdata, test_png_BytesIO):
     mocked_resource = Mock()
     with patch("boto3.client", Mock(return_value=mocked_connection)):
         with patch("boto3.resource", Mock(return_value=mocked_resource)):
-            OpenOversight.app.utils.upload_obj_to_s3(test_png_BytesIO, "test_cop1.png")
+            OpenOversight.app.util.utils.upload_obj_to_s3(test_png_BytesIO, "test_cop1.png")
 
     assert (
         mocked_connection.method_calls[0][2]["ExtraArgs"]["ContentType"] == "image/png"
@@ -195,7 +195,7 @@ def test_s3_upload_jpeg(mockdata, test_jpg_BytesIO):
     mocked_resource = Mock()
     with patch("boto3.client", Mock(return_value=mocked_connection)):
         with patch("boto3.resource", Mock(return_value=mocked_resource)):
-            OpenOversight.app.utils.upload_obj_to_s3(test_jpg_BytesIO, "test_cop5.jpg")
+            OpenOversight.app.util.utils.upload_obj_to_s3(test_jpg_BytesIO, "test_cop5.jpg")
 
     assert (
         mocked_connection.method_calls[0][2]["ExtraArgs"]["ContentType"] == "image/jpeg"
@@ -210,21 +210,21 @@ def test_user_can_submit_allowed_file(mockdata):
         "valid_photo.PNG",
         "valid_photo.JPG",
     ]:
-        assert OpenOversight.app.utils.allowed_file(file_to_submit) is True
+        assert OpenOversight.app.util.utils.allowed_file(file_to_submit) is True
 
 
 def test_user_cannot_submit_malicious_file(mockdata):
     file_to_submit = "passwd"
-    assert OpenOversight.app.utils.allowed_file(file_to_submit) is False
+    assert OpenOversight.app.util.utils.allowed_file(file_to_submit) is False
 
 
 def test_user_cannot_submit_invalid_file_extension(mockdata):
     file_to_submit = "tests/test_models.py"
-    assert OpenOversight.app.utils.allowed_file(file_to_submit) is False
+    assert OpenOversight.app.util.utils.allowed_file(file_to_submit) is False
 
 
 def test_unit_choices(mockdata):
-    unit_choices = [str(x) for x in OpenOversight.app.utils.unit_choices()]
+    unit_choices = [str(x) for x in OpenOversight.app.util.utils.unit_choices()]
     assert "Unit: Bureau of Organized Crime" in unit_choices
